separate key from util
