+++ hooks/queries/keys.ts
@@ -1,0 +1,5 @@
+import { createQueryKeys } from '@/utils/createQueryKeys';
+
+export const MEMBER_KEY = createQueryKeys('members');
+export const GROUP_KEY = createQueryKeys('groups');
+export const INVITATION_KEY = createQueryKeys('invitations');
